+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.16] - 2023-04-12
+
+### Fixed
+
+- `fieldProxy` and `formFieldProxy` are now exported from `sveltekit-superforms/client`.
+
 ## [0.6.15] - 2023-04-12
 
 ### Fixed